+++ library/LocalRegions/NodalTriExp.h
@@ -228,14 +228,6 @@ private:
         IProductWRTDerivBase(dir, inarray, outarray);
     }
 
-    void v_StdPhysDeriv(
-        const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &out_d0, Array<OneD, NekDouble> &out_d1,
-        Array<OneD, NekDouble> &out_d2 = NullNekDouble1DArray) override
-    {
-        StdTriExp::v_PhysDeriv(inarray, out_d0, out_d1, out_d2);
-    }
-
     void v_PhysDeriv(const Array<OneD, const NekDouble> &inarray,
                      Array<OneD, NekDouble> &out_d0,
                      Array<OneD, NekDouble> &out_d1,