+++ library/LocalRegions/SegExp.cpp
@@ -960,6 +960,14 @@ void SegExp::v_LaplacianMatrixOp(
     v_IProductWRTBase(m_base[0]->GetDbdata(), dPhysValuesdx, outarray, 1);
 }
 
+void SegExp::v_LaplacianMatrixOp(const int k1, const int k2,
+                                 const Array<OneD, const NekDouble> &inarray,
+                                 Array<OneD, NekDouble> &outarray,
+                                 const StdRegions::StdMatrixKey &mkey)
+{
+    StdExpansion::LaplacianMatrixOp_MatFree(k1, k2, inarray, outarray, mkey);
+}
+
 void SegExp::v_HelmholtzMatrixOp(const Array<OneD, const NekDouble> &inarray,
                                  Array<OneD, NekDouble> &outarray,
                                  const StdRegions::StdMatrixKey &mkey)